Ajout des surfaces + problème collision brique

+++ surfaceNormale.h
@@ -8,7 +8,7 @@ class surfaceNormale : public surface
     public:
         surfaceNormale();
         virtual ~surfaceNormale();
-        virtual int getVitesse() const override;
+        virtual double getVitesse() const override;
         virtual bool getMorte() const override;
     private:
 };
